requirements.txt
pymysql
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-rest-swagger
django-model-utils